Move Writer class to the correct namespace
