+++ SAGESharp.Tests/IO/Binary/WriterTests.cs
@@ -6,12 +6,12 @@
 using NSubstitute;
 using NSubstitute.ClearExtensions;
 using NUnit.Framework;
-using SAGESharp.IO;
+using SAGESharp.IO.Binary;
 using System;
 
-namespace SAGESharp.Tests.IO
+namespace SAGESharp.Tests.IO.Binary
 {
-    class BinaryWriterTests
+    class WriterTests
     {
         private readonly BinaryWriterSubstitute binaryWriter = BinaryWriterSubstitute.New();
